updates, experiments

+++ code/run.py
@@ -59,6 +59,12 @@ class MetricsLogger(Callback):
         self.csv_path = csv_path
         self.fold_number = fold_number
         self.base_name = base_name
+        # Check if file exists to decide whether to write headers
+        if not os.path.exists(csv_path):
+            with open(csv_path, 'w', newline='') as csvfile:
+                writer = csv.writer(csvfile)
+                # Write the header
+                writer.writerow(['Base Name', 'Epoch', 'Fold', 'Loss', 'Accuracy', 'Validation Loss', 'Validation Accuracy'])
 
     def on_epoch_end(self, epoch, logs=None):
         logs = logs or {}
@@ -103,6 +109,8 @@ num_classes = config.num_classes
 buffer_size = config.buffer_size
 ltn_batch = config.ltn_batch
 results_path = config.results_path
+S = config.S
+processed_file_tracker = config.processed_file_tracker
 
 # LTN metrics and groundings
 metrics_dict = {
@@ -116,8 +124,8 @@ Not = ltn.Wrapper_Connective(ltn.fuzzy_ops.Not_Std())
 And = ltn.Wrapper_Connective(ltn.fuzzy_ops.And_Prod())
 Or = ltn.Wrapper_Connective(ltn.fuzzy_ops.Or_ProbSum())
 Implies = ltn.Wrapper_Connective(ltn.fuzzy_ops.Implies_Reichenbach())
-Forall = ltn.Wrapper_Quantifier(ltn.fuzzy_ops.Aggreg_pMeanError(p=4),semantics="forall")
-formula_aggregator = ltn.Wrapper_Formula_Aggregator(ltn.fuzzy_ops.Aggreg_pMeanError(p=4))
+Forall = ltn.Wrapper_Quantifier(ltn.fuzzy_ops.Aggreg_pMeanError(p=3),semantics="forall")
+formula_aggregator = ltn.Wrapper_Formula_Aggregator(ltn.fuzzy_ops.Aggreg_pMeanError(p=3))
 
 class_0 = ltn.Constant(0, trainable=False)
 class_1 = ltn.Constant(1, trainable=False)
@@ -129,9 +137,12 @@ class_6 = ltn.Constant(6, trainable=False)
 class_7 = ltn.Constant(7, trainable=False)
 class_8 = ltn.Constant(8, trainable=False)
 
-with open(results_path, 'w', newline='') as csvfile:
-    writer = csv.writer(csvfile)
-    writer.writerow(['Speed','Epoch', 'Fold', 'Loss', 'Accuracy', 'Validation Loss', 'Validation Accuracy'])
+# Check if the file exists to decide whether to write headers
+if not os.path.exists(results_path):
+    with open(results_path, 'w', newline='') as csvfile:
+        writer = csv.writer(csvfile)
+        # Write the header
+        writer.writerow(['Speed', 'Epoch', 'Fold', 'Loss', 'Accuracy', 'Validation Loss', 'Validation Accuracy'])
 
 
 
@@ -181,13 +192,23 @@ def main():
         counter = 0
         console = Console()
         processed_bases = set()
+        
+        if os.path.exists(processed_file_tracker):
+            with open(processed_file_tracker, "r") as file:
+                processed_bases = set(file.read().splitlines())
+        
         metrics_summary = []
 
         for file in sorted(os.listdir(sequences_directory)):
             if "_train_scaled_sequences.npy" in file:
+                
+                if counter >= S:
+                    break
+                
                 base_name = file.replace("_train_scaled_sequences.npy", "")
                 if base_name in processed_bases:
                     continue
+                processed_bases.add(base_name)
                 counter+=1
                 
                 
@@ -261,11 +282,18 @@ def main():
 
                     console.print(f"Fold {fold+1} Accuracy: {accuracy:.4f}, Precision: {precision:.4f}, Recall: {recall:.4f}, F1: {f1:.4f}")
                     
-                    #class_importances = permutation_importance_per_class(model, X_val_fold, y_val_fold, n_repeats=4, n_samples=n_samples)
-                    #for class_id, importances in class_importances.items():
-                        #print(f"{class_id} Feature Importances:", importances)
+                    class_importances = permutation_importance_per_class(model, X_val_fold, y_val_fold, n_repeats=4, n_samples=n_samples)
+                    for class_id, importances in class_importances.items():
+                        print(f"{class_id} Feature Importances:", importances)
                     
+                    importances_matrix = np.array(list(class_importances.values()))
+                    # Calculate the average importance across all classes
+                    average_importances = np.mean(importances_matrix, axis=0)
                     
+                    normalized_average_importances = normalize_importances(average_importances)
+
+                    print("Normalized Average Feature Importances:", normalized_average_importances)
+
                     p = ltn.Predicate.FromLogits(model, activation_function="softmax", with_class_indexing=True)
                     
                     @tf.function
@@ -293,7 +321,7 @@ def main():
                         sat_level = formula_aggregator(axioms).tensor
                         return sat_level
 
-                    optimizer = tf.keras.optimizers.Adam(learning_rate=learning_rate / 10)
+                    optimizer = tf.keras.optimizers.Adam(learning_rate=learning_rate)
                     
                     @tf.function
                     def train_step(features, labels):
@@ -318,14 +346,12 @@ def main():
                         predictions = model([features])
                         metrics_dict['test_accuracy'](tf.one_hot(labels,9),predictions)
                     
-                    #normalized_importances = {}
-
-                    # for class_label, importances in class_importances.items():
-                    #     normalized = normalize_importances(importances)
-                    #     normalized_importances[class_label] = normalized
+                    
+                    X_train_fold_weighted = X_train_fold * np.array(normalized_average_importances)
+                    X_val_fold_weighted = X_val_fold * np.array(normalized_average_importances)
                         
-                    ds_train_fold = tf.data.Dataset.from_tensor_slices((X_train_fold, y_train_fold))
-                    ds_val_fold = tf.data.Dataset.from_tensor_slices((X_val_fold, y_val_fold))
+                    ds_train_fold = tf.data.Dataset.from_tensor_slices((X_train_fold_weighted, y_train_fold))
+                    ds_val_fold = tf.data.Dataset.from_tensor_slices((X_val_fold_weighted, y_val_fold))
                     
                     ds_train_fold = ds_train_fold.batch(batch_size)
                     ds_val_fold = ds_val_fold.batch(batch_size)
@@ -371,8 +397,8 @@ def main():
 
 
                 
-                if counter>2:
-                    break
+                # if counter>2:
+                #     break
             
         concatenate_and_delete_ltn_csv_files(results_path_ltn, "results/results_ltn.csv")
 
@@ -387,6 +413,17 @@ def main():
         console.print(f"Overall Average Precision: {overall_avg_precision:.4f}")
         console.print(f"Overall Average Recall: {overall_avg_recall:.4f}")
         console.print(f"Overall Average F1: {overall_avg_f1:.4f}")
+        
+        # After the loop, save the updated processed bases to file
+        with open(processed_file_tracker, "w") as file:
+            for base in processed_bases:
+                file.write(base + "\n")
+        
+        
+
+        
+                
+
